refactor(account): restructure code and fix bugs, use new service to guard images

+++ src/account/services/update-project.service.ts
@@ -1,9 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { S3ImageManagementService } from '../../aws/s3/services/s3-image-management.service';
 import { ImageRepository } from '../../database/repositories/image.repository';
 import { ProjectRepository } from '../../database/repositories/project.repository';
 import { UpdatedProject } from '../models/updated-project.model';
 import { Types } from 'mongoose';
+import { ImageAssociationService } from '../../utils/services/image-association.service';
+import { Project } from '../../database/schemas/project.schema';
 
 @Injectable()
 export class UpdateProjectService {
@@ -11,10 +13,11 @@ export class UpdateProjectService {
         private projectRepository: ProjectRepository,
         private imageRepository: ImageRepository,
         private s3ImageManagementService: S3ImageManagementService,
+        private imageAssociationService: ImageAssociationService,
     ) {}
 
     public async execute(updatedProject: UpdatedProject): Promise<void> {
-        const dbProject = await this.projectRepository.findOneRaw({
+        let dbProject = await this.projectRepository.findOneRaw({
             find: { id: updatedProject.id },
         });
 
@@ -27,60 +30,90 @@ export class UpdateProjectService {
         dbProject.category = updatedProject.category ?? dbProject.category;
         dbProject.socialMedia = updatedProject.socialMedia ?? dbProject.socialMedia;
 
+        dbProject = await this.updateLogo(dbProject, updatedProject);
+        dbProject = await this.updateImages(dbProject, updatedProject);
+
+        await this.projectRepository.update(dbProject);
+    }
+
+    private async updateLogo(dbProject: Project, updatedProject: UpdatedProject): Promise<Project> {
         if (updatedProject.logo) {
-            // Delete old logo
             if (dbProject.logo) {
                 const oldLogo = await this.imageRepository.findOneRaw({
-                    find: { _id: dbProject.logo },
+                    find: { _id: dbProject.logo as Types.ObjectId },
                 });
 
                 await this.s3ImageManagementService.deleteFileOnS3(oldLogo);
                 await this.imageRepository.delete({ find: { _id: oldLogo._id } });
-                dbProject.logo = undefined;
+                dbProject.logo = null;
             }
 
-            // Set new logo - if present
-            if (updatedProject.logo.id) {
-                // TODO: We need to verify that the image has not been used yet
+            if (Object.keys(updatedProject.logo).length !== 0) {
                 const newLogo = await this.imageRepository.findOneRaw({
                     find: { id: updatedProject.logo.id },
                 });
+
+                if (!(await this.imageAssociationService.isImageUnassociated(newLogo))) {
+                    throw new UnauthorizedException(
+                        'New logo already belongs to a different project',
+                    );
+                }
 
                 dbProject.logo = newLogo._id;
             }
         }
 
+        return dbProject;
+    }
+
+    private async updateImages(
+        dbProject: Project,
+        updatedProject: UpdatedProject,
+    ): Promise<Project> {
         if (updatedProject.images) {
             dbProject.images = (dbProject.images as Types.ObjectId[]) ?? [];
-            const oldImageIds = [...dbProject.images];
+            const oldImages = [...dbProject.images] as Types.ObjectId[];
+            const newImages = [] as Types.ObjectId[];
 
             for (const image of updatedProject.images) {
-                const imageModel = await this.imageRepository.findOneRaw({
+                const newImage = await this.imageRepository.findOneRaw({
                     find: { id: image.id },
                 });
-                const index = oldImageIds.findIndex((id) => id.equals(imageModel._id));
 
-                if (index > -1) {
-                    // Existing image and we want to keep it
-                    oldImageIds.splice(index, 1);
-                } else {
-                    // New image
-                    // TODO: We need to verify that the image has not been used yet
-                    dbProject.images.push(imageModel._id);
+                if (
+                    !(await this.imageAssociationService.isImageUnassociated(newImage)) &&
+                    !(await this.imageAssociationService.isProjectPictureOwner(dbProject, newImage))
+                ) {
+                    throw new UnauthorizedException(
+                        'New image already belongs to a different project',
+                    );
                 }
+
+                newImages.push(newImage._id);
             }
 
-            for (const id of oldImageIds) {
+            for (const oldImageId of oldImages) {
+                let deleteOldImage: boolean = true;
+                for (const newImageId of newImages) {
+                    if (newImageId.toString() === oldImageId.toString()) {
+                        deleteOldImage = false;
+                    }
+                }
+
+                if (!deleteOldImage) {
+                    continue;
+                }
+
                 const imageModel = await this.imageRepository.findOneRaw({
-                    find: { _id: id },
+                    find: { _id: oldImageId },
                 });
                 await this.s3ImageManagementService.deleteFileOnS3(imageModel);
-                await this.imageRepository.delete({ find: { _id: id } });
-                const index = dbProject.images.findIndex((el) => el.equals(id));
-                dbProject.images.splice(index, 1);
+                await this.imageRepository.delete({ find: { _id: oldImageId } });
             }
+
+            dbProject.images = newImages;
         }
 
-        await this.projectRepository.update(dbProject);
+        return dbProject;
     }
 }
